refactor(familiars): isolate chat capability surface
